chore: update GitHub Actions to latest versions (#2)
* chore: update GitHub Actions to latest versions

Upgrade GitHub Actions in the Unity build workflow to 
the latest versions. This includes updating the checkout, 
cache, and upload-artifact actions to v3 for improved 
performance and security.

* chore: update upload-artifact action to v4

Upgrade the GitHub Actions upload-artifact action from v3 to v4 
to leverage the latest features and improvements. This change 
ensures better performance and compatibility with future updates.