feat: implement unit spawning system and database
Add a new UnitSpawner class to handle unit instantiation 
based on data from a UnitDatabase. Update PlayerController 
to utilize the new spawning method. Create UnitData 
scriptable object to define unit properties. Add 
zombie and player assets to the database for easy 
management and instantiation. This refactor improves 
the modularity and scalability of unit management in 
the game.